requirements.txt
streamlit>=1.33
pydantic>=2.7
jinja2>=3.1
httpx>=0.27
python-dotenv>=1.0
openai>=1.40
tenacity>=8.2
pytest>=8.2
pytest-mock>=3.14
streamlit-sortables>=0.2.0
PyYAML>=6.0

jsonschema>=4.0
validators>=0.20
google-cloud-storage>=2.16
